resolving a bug where my report was not being downloaded by the user

+++ src/main/java/com/meubovinoapp/dao/ReportDAO.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ReportDAO extends JpaRepository<Report,Long> {
+public interface ReportDAO extends JpaRepository<Report,Integer> {
 
     List<Report> getAllReports();
